% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable containing the REDCap API token.}

\item{redcapUrl}{The URL to the REDCap API endpoint.}

\item{redcapReportId}{The ID of the REDCap report to download.}
}
\value{
A tibble containing the data from the REDCap report.
}
\description{
This function downloads a specified report from REDCap using API credentials stored in an environment variable.
The result is returned as a tibble.
}
