% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the maximum likelihood estimate of parameter p for Bernoulli distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector of binary data (0s and 1s).}
}
\value{
A list containing the maximum log-likelihood and the corresponding parameter p.
}
\description{
Calculate the maximum likelihood estimate of parameter p for Bernoulli distribution
}
\examples{
data = c(1, 0, 0, 0, 1, 1, 1)
logLikBernoulli(data)
}
