% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Approximate data matrix using a specified number of principal components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or data frame of data to approximate.}

\item{npc}{The number of principal components to use for the approximation.}
}
\value{
A numeric matrix approximating the original data, rescaled and recentered.
}
\description{
This function performs PCA on a data matrix \code{x} and uses \code{npc} principal components to
approximate and then rescale and recenter the data to match the original data's statistics.
}
\examples{
data_matrix = matrix(rnorm(100), nrow=10)
approx_data = pcApprox(data_matrix, 2)
}
