% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize variable names in a tibble}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble whose column names are to be standardized.}
}
\value{
A tibble with standardized column names in "small_camel" case.
}
\description{
This function uses \code{dplyr::rename_with} and \code{janitor::make_clean_names} to clean and standardize the
column names of a tibble. It then converts these names to "small_camel" case using \code{snakecase::to_small_camel_case}.
}
\examples{
library(tibble)
data = tibble(`some data` = 1:5, `!!more data!!` = 5:1)
standardized_data = standardizeNames(data)
}
