% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Calculate and plot a survival curve S(t)}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numerical vector indicating whether an event occurred (1) or was censored (0).}

\item{time}{Numerical vector indicating the time at which each event or censorship occurred.}
}
\value{
A ggplot object showing the survival curve.
}
\description{
This function takes numerical vectors for status (1 for events, 0 for censored) and time,
and calculates and plots a survival curve using a step function. It does not use any specialized
survival analysis packages, and computes the Kaplan-Meier estimator manually.
}
\examples{
data = read.csv("https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv")
survCurv(data$status, data$time)
}
