% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscaled.R
\name{unscale}
\alias{unscale}
\title{Reverse the effects of scaling on a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been scaled.}
}
\value{
A numeric vector with the scaling reversed.
}
\description{
This function takes a scaled vector \code{x} and reverses the scaling process. It assumes that \code{x}
has been scaled using the base R \code{scale()} function, which stores the center (mean) and scale (standard deviation)
as attributes of the returned object.
}
\examples{
original = 1:10
scaled = scale(original)
unscaled = unscale(scaled)
all.equal(unscaled, original) # Should return TRUE
}
